refactor(editor): extract Emit.GUI module for structured chrome data sends
Move all 9 send_gui_* functions and their business logic helpers from
Emit into Emit.GUI with a single public sync_chrome/1 entry point.
Capabilities.gui? check consolidated to one call site in Emit.emit/2.

Emit.ex: 690 → 452 lines, now focused on frame-to-commands conversion.

New module:
- lib/minga/editor/render_pipeline/emit/gui.ex (266 lines)

Business logic helpers moved: build_status_bar_data, build_agent_chat_data,
current_cursor_screen_pos, resolve_git_branch, active_buffer_path.

Closes #691
